refactoring first pass
