+++ setup.py
@@ -54,7 +54,7 @@ setuptools.setup(
     },
     entry_points={
         'console_scripts': [
-            'eva = eva.eva_base:main',
+            'eva = eva.eva_driver:main',
             'eva_tests = eva.eva_tests:main',
         ],
     },
