+++ src/eva/eva_base.py
@@ -103,126 +103,3 @@ class EvaFactory():
         # Return implementation of the class (calls base class constructor that is above)
         # -------------------------------------------------------------------------------
         return eva_class(eva_class_name, config, eva_logger, timing)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-def eva(eva_config, eva_logger=None):
-
-    # Create timing object
-    timing = Timing()
-
-    # Create temporary logger
-    logger = Logger('EvaSetup')
-
-    logger.info('Starting Eva')
-
-    # Convert incoming config (either dictionary or file) to dictionary
-    timing.start('Generate Dictionary')
-    if isinstance(eva_config, dict):
-        eva_dict = eva_config
-    else:
-        # Create dictionary from the input file
-        eva_dict = load_yaml_file(eva_config, logger)
-
-    # Get the list of applications
-    try:
-        diagnostic_configs = eva_dict['diagnostics']
-    except KeyError:
-        logger.abort('eva configuration must contain \'diagnostics\' and it should provide a ' +
-                     'list of diagnostics to be run.')
-
-    if not isinstance(diagnostic_configs, list):
-        raise TypeError(f'diagnostics should be a list, it was type: {type(diagnostic_configs)}')
-    timing.stop('Generate Dictionary')
-
-    # Loop over the applications and run
-    for diagnostic_config in diagnostic_configs:
-
-        # Each diagnostic should have two dictionaries: data and graphics
-        if not all(sub_config in diagnostic_config for sub_config in ['data', 'graphics']):
-            msg = "diagnostic config must contain 'data' and 'graphics'"
-            raise KeyError(msg)
-
-        # Create the data collections
-        # ---------------------------
-        data_collections = DataCollections()
-
-        # Create the data object
-        creator = EvaFactory()
-        timing.start('DataConstructor')
-        eva_data_object = creator.create_eva_object('DataDriver',
-                                                    'data',
-                                                    diagnostic_config,
-                                                    eva_logger,
-                                                    timing)
-        timing.stop('DataConstructor')
-
-        # Prepare diagnostic data
-        logger.info(f'Running execute for {eva_data_object.name}')
-        timing.start('DataExecute')
-        eva_data_object.execute(data_collections, timing)
-        timing.stop('DataExecute')
-
-        # Create the transforms
-        if 'transforms' in diagnostic_config:
-            timing.start('TransformsConstructor')
-            eva_transform_object = creator.create_eva_object('TransformDriver',
-                                                             'transforms',
-                                                             diagnostic_config,
-                                                             eva_logger,
-                                                             timing)
-            timing.stop('TransformsConstructor')
-            logger.info(f'Running execute for {eva_transform_object.name}')
-            timing.start('TransformsExecute')
-            eva_transform_object.execute(data_collections, timing)
-            timing.stop('TransformsExecute')
-
-        # Create the figure object
-        timing.start('FigureDriverConstructor')
-        eva_figure_object = creator.create_eva_object('FigureDriver',
-                                                      'plot_tools',
-                                                      diagnostic_config,
-                                                      eva_logger,
-                                                      timing)
-        timing.stop('FigureDriverConstructor')
-
-        # Generate figure(s)
-        logger.info(f'Running execute for {eva_figure_object.name}')
-        timing.start('FigureDriverExecute')
-        eva_figure_object.execute(data_collections, timing)
-        timing.stop('FigureDriverExecute')
-
-    timing.finalize()
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-def main():
-
-    # Arguments
-    # ---------
-    parser = argparse.ArgumentParser()
-    parser.add_argument('config_file', type=str, help='Configuration YAML file for driving ' +
-                        'the diagnostic. See documentation/examples for how to configure the YAML.')
-
-    # Get the configuation file
-    args = parser.parse_args()
-    config_file = args.config_file
-
-    assert os.path.exists(config_file), "File " + config_file + " not found"
-
-    # Run the diagnostic(s)
-    eva(config_file)
-
-
-# --------------------------------------------------------------------------------------------------
-
-
-if __name__ == "__main__":
-    main()
-
-
-# --------------------------------------------------------------------------------------------------
